requirements.txt
PyQt6
volatility3
pefile
requests
reportlab
matplotlib
cryptography
pdfminer.six
